requirements.txt
streamlit>=1.28
requests
pillow
opencv-python-headless
pandas
